# backend/codeblurb-backend/src/main/java/hu/codeblurb/backend/controller/dto/IncorrectSolutionsResponse.java
package hu.codeblurb.backend.controller.dto;

public record IncorrectSolutionsResponse(Integer incorrectQuestionId, Integer incorrectSolutionSent,
                                         Integer correctSolution) {
}