# backend/codeblurb-backend/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
paths:
...
  /content/quiz/solution/{contentId}:
    post:
      tags:
        - Content
      operationId: checkSolutionForQuiz
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizSolutionRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/QuizSolutionResponse'
  /content/code/solution/{contentId}:
    post:
      tags:
        - Content
      operationId: runSolutionFor
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/CodeSolutionResponse'
...
components:
  schemas:
...
    QuizSolutionRequest:
      type: object
      properties:
        solutions:
          type: object
          additionalProperties:
            type: integer
            format: int32
    IncorrectSolutionsResponse:
      type: object
      properties:
        incorrectQuestionId:
          type: integer
          format: int32
        incorrectSolutionSent:
          type: integer
          format: int32
        correctSolution:
          type: integer
          format: int32
    QuizSolutionResponse:
      type: object
      properties:
        correctAnswerQuestionIds:
          type: array
          items:
            type: integer
            format: int32
        incorrectSolutions:
          type: array
          items:
            $ref: '#/components/schemas/IncorrectSolutionsResponse'
    CodeSolutionResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/TestCaseResultResponse'
        overallResult:
          type: string
          enum:
            - ALL_PASSED
            - TESTCASES_FAILED
    TestCaseResultResponse:
      type: object
      properties:
        expected:
          type: string
        actual:
          type: string
        result:
          type: string
          enum:
            - PASSED
            - FAILED
        hints:
          type: array
          items:
            type: string
...